// client/src/app/models/socket-actions.ts
export enum ACTIONS {
  JOIN = 'join',
  LEAVE = 'leave',
  MESSAGE = 'message',
  SHARE_ROOMS = 'share-rooms',
  ADD_PEER = 'add-peer',
  REMOVE_PEER = 'remove-peer',
  RELAY_SDP = 'relay-sdp',
  RELAY_ICE = 'relay-ice',
  ICE_CANDIDATE = 'ice-candidate',
  SESSION_DESCRIPTION = 'session-description'
}